Work-in-progress commit
Work is going on to this into a stand-alone probe station app
To do this the code needs to run on more machines and have tight
version control. To do this, I simply commit the entire thing now -
once the application is finished, this code will move entirely to
PyExpLabSys.